backend/requirements.txt
Django>=5.2.0
djangorestframework>=3.16.0
djangorestframework-simplejwt>=5.5.0
django-cors-headers>=4.9.0
django-csp>=3.8
pyodbc>=5.0.0
